+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 
 ## [Unreleased]
+### Changed
+- The 'current hour' counter in at the bottom now displays 1-24, instead of 0-23
 
 ## [0.6.1] - 2019-04-09
 ### Added